Create repository log handler
