+++ src/core/domain/codeReviewSettingsLog/contracts/codeReviewSettingsLog.service.contract.ts
@@ -1,7 +1,8 @@
 import { ICodeReviewSettingsLogRepository } from './codeReviewSettingsLog.repository.contract';
 import { KodyRuleLogParams } from '@/core/infrastructure/adapters/services/codeReviewSettingsLog/kodyRulesLog.handler';
 import { CodeReviewConfigLogParams } from '@/core/infrastructure/adapters/services/codeReviewSettingsLog/codeReviewConfigLog.handler';
-import { RepositoriesLogParams, IntegrationLogParams, UserStatusLogParams } from '@/core/infrastructure/adapters/services/codeReviewSettingsLog/types/logParams.types';
+import { IntegrationLogParams, UserStatusLogParams } from '@/core/infrastructure/adapters/services/codeReviewSettingsLog/types/logParams.types';
+import { RepositoriesLogParams } from '@/core/infrastructure/adapters/services/codeReviewSettingsLog/repositoriesLog.handler';
 
 export const CODE_REVIEW_SETTINGS_LOG_SERVICE_TOKEN = Symbol(
     'CodeReviewSettingsLogService',
