+++ src/core/infrastructure/adapters/services/codeReviewSettingsLog/codeReviewSettingsLog.service.ts
@@ -11,13 +11,16 @@ import {
     CodeReviewConfigLogHandler,
     CodeReviewConfigLogParams,
 } from './codeReviewConfigLog.handler';
+import {
+    RepositoriesLogHandler,
+    RepositoriesLogParams,
+} from './repositoriesLog.handler';
 import { UnifiedLogHandler } from './unifiedLog.handler';
 import {
     ActionType,
     ConfigLevel,
 } from '@/config/types/general/codeReviewSettingsLog.type';
 import {
-    RepositoriesLogParams,
     IntegrationLogParams,
     UserStatusLogParams,
 } from './types/logParams.types';
@@ -42,6 +45,8 @@ export class CodeReviewSettingsLogService
         private readonly kodyRulesLogHandler: KodyRulesLogHandler,
 
         private readonly codeReviewConfigLogHandler: CodeReviewConfigLogHandler,
+
+        private readonly repositoriesLogHandler: RepositoriesLogHandler,
     ) {}
 
     async create(
@@ -73,116 +78,7 @@ export class CodeReviewSettingsLogService
     public async registerRepositoriesLog(
         params: RepositoriesLogParams,
     ): Promise<void> {
-        const {
-            organizationAndTeamData,
-            userInfo,
-            actionType,
-            addedRepositories = [],
-            removedRepositories = [],
-            sourceRepository,
-            targetRepository,
-        } = params;
-
-        let changedData: ChangedDataToExport[] = [];
-        let finalActionType: ActionType;
-        let configLevel: ConfigLevel;
-        let repository: { id: string; name: string } | undefined;
-
-        // ✅ Decidir pela action type
-        if (
-            actionType === ActionType.ADD &&
-            sourceRepository &&
-            targetRepository
-        ) {
-            // COPY operation (individual)
-            const isSourceGlobal = sourceRepository.id === 'global';
-            const sourceName = isSourceGlobal
-                ? 'Global Settings'
-                : sourceRepository.name;
-
-            changedData = [
-                {
-                    actionDescription: 'Repository Configuration Copied',
-                    previousValue: null,
-                    currentValue: {
-                        sourceRepository: {
-                            id: sourceRepository.id,
-                            name: sourceName,
-                            isGlobal: isSourceGlobal,
-                        },
-                        targetRepository: {
-                            id: targetRepository.id,
-                            name: targetRepository.name,
-                        },
-                    },
-                    description: `User ${userInfo.userEmail} copied code review configuration from ${isSourceGlobal ? 'Global Settings' : `"${sourceName}"`} to repository "${targetRepository.name}"`,
-                },
-            ];
-
-            finalActionType = ActionType.ADD;
-            configLevel = ConfigLevel.REPOSITORY;
-            repository = {
-                id: targetRepository.id,
-                name: targetRepository.name,
-            };
-        } else {
-            if (
-                addedRepositories.length === 0 &&
-                removedRepositories.length === 0
-            ) {
-                return;
-            }
-
-            const allChangedData: ChangedDataToExport[] = [];
-
-            // Adicionar changedData para cada repo adicionado
-            addedRepositories.forEach((repo) => {
-                allChangedData.push({
-                    actionDescription: 'Repository Added',
-                    previousValue: null,
-                    currentValue: {
-                        id: repo.id,
-                        name: repo.name,
-                    },
-                    description: `User ${userInfo.userEmail} added repository "${repo.name}" to code review settings`,
-                });
-            });
-
-            // Adicionar changedData para cada repo removido
-            removedRepositories.forEach((repo) => {
-                allChangedData.push({
-                    actionDescription: 'Repository Removed',
-                    previousValue: {
-                        id: repo.id,
-                        name: repo.name,
-                    },
-                    currentValue: null,
-                    description: `User ${userInfo.userEmail} removed repository "${repo.name}" from code review settings`,
-                });
-            });
-
-            changedData = allChangedData;
-
-            finalActionType = actionType;
-            configLevel = ConfigLevel.GLOBAL;
-            repository = undefined;
-        }
-
-        // Salvar um registro só com todos os changedData
-        await this.codeReviewSettingsLogRepository.create({
-            organizationId: organizationAndTeamData.organizationId,
-            teamId: organizationAndTeamData.teamId,
-            action: finalActionType,
-            userInfo: {
-                userId: userInfo.userId,
-                userEmail: userInfo.userEmail,
-            },
-            changeMetadata: {
-                configLevel,
-                repository,
-            },
-            changedData,
-        });
+        await this.repositoriesLogHandler.logRepositoriesAction(params);
     }
 
     public async registerIntegrationLog(
