+++ src/core/infrastructure/adapters/services/codeReviewSettingsLog/types/logParams.types.ts
@@ -1,16 +1,6 @@
 import { OrganizationAndTeamData } from '@/config/types/general/organizationAndTeamData';
 import { ActionType, UserInfo } from '@/config/types/general/codeReviewSettingsLog.type';
 
-export interface RepositoriesLogParams {
-    organizationAndTeamData: OrganizationAndTeamData;
-    userInfo: UserInfo;
-    actionType: ActionType;
-    addedRepositories?: Array<{ id: string; name: string }>;
-    removedRepositories?: Array<{ id: string; name: string }>;
-    sourceRepository?: { id: string; name: string };
-    targetRepository?: { id: string; name: string };
-}
-
 export interface IntegrationLogParams {
     organizationAndTeamData: OrganizationAndTeamData;
     userInfo: UserInfo;
